# Conflicts:
#	app/src/main/java/com/bink/wallet/utils/Bindings.kt

+++ app/src/main/java/com/bink/wallet/model/response/membership_card/MembershipTransactions.kt
@@ -8,6 +8,6 @@ import kotlinx.android.parcel.Parcelize
 @JsonClass(generateAdapter = true)
 data class MembershipTransactions (
 	val description : String?,
-	val timestamp : Int?,
+	val timestamp : Long?,
 	val amounts : List<Amounts>?
 ) : Parcelable
